requirements.txt
Flask==3.0.0
groq==0.11.0
python-dotenv==1.0.0
Werkzeug==3.0.1
httpx==0.27.0

# Production server
gunicorn==21.2.0

# Optional: For better performance
gevent==24.2.1
